Fix hack by adding another hack
In vim, when one did ,gg' and then hit enter, it was not actually
calling CancelWindowCommandAction, it was just typing the command that
should have done so and hitting enter. Without actually calling it. Not
sure why, but forcibly returning to normal mode first seems to make it
work. Very strange. Wish I understood why this fixed it. It's midnight.